Models atualizados: User herda de IdentityUser Controllers criados Base de dados criada

+++ PlantsRPetsProjeto.Server/Models/Metric.cs
@@ -5,7 +5,7 @@
         public int MetricId { get; set; }
         public int UserId { get; set; }
         public int TotalPlants { get; set; }
-        public decimal WaterSaved { get; set; }
-        public decimal CarbonFootprintReduction { get; set; }
+        public double WaterSaved { get; set; }
+        public double CarbonFootprintReduction { get; set; }
     }
 }
